minimal/requirements.txt
fastapi
uvicorn[standard]
requests
python-jose[cryptography]
python-dotenv